refactor: split NewGraphics ideas into modules
