+++ src/NewGraphics/Loading/ILoadsResourceTable.cs
@@ -1,0 +1,12 @@
+using RenderMaster.src.NewGraphics.Resources;
+
+namespace RenderMaster.src.NewGraphics.Loading
+{
+    interface ILoadsResourceTable
+    {
+        CPUResourceTable ResourceTable { get; }
+
+        void LoadResources(CPUResourceTable table);
+    }
+}
+
