+++ src/NewGraphics/Scene/SceneNode.cs
@@ -1,0 +1,15 @@
+using RenderMaster.src.NewGraphics.Resources;
+using RenderMaster.src.NewGraphics.Types;
+
+namespace RenderMaster.src.NewGraphics.Scene
+{
+    using MeshHandle = Handle<PreparedMeshBuffer>;
+    using MaterialHandle = Handle<MaterialCPU>;
+
+    class SceneNode
+    {
+        public required MeshHandle mesh { get; init; }
+        public required MaterialHandle material { get; init; }
+    }
+}
+
