+++ src/NewGraphics/Types/Handle.cs
@@ -1,0 +1,5 @@
+namespace RenderMaster.src.NewGraphics.Types
+{
+    readonly record struct Handle<T>(int Id);
+}
+
